requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.9.1.post1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
av==10.0.0
blinker==1.7.0
certifi==2023.5.7
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
distro==1.8.0
djitellopy==2.5.0
dm-tree==0.1.8
dnspython==2.6.1
droppy==1.0.0b0
email_validator==2.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.1
flatbuffers==24.3.7
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.10.0
future==1.0.0
gast==0.5.4
google-images-download @ git+https://github.com/Joeclinton1/google-images-download.git@e91e6a38ad654877f59edc7894822f5319d75df1
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.62.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huepy==1.2.1
huggingface-hub==0.22.2
idna==3.4
imageio==2.34.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instabot==0.117.0
instagrapi==2.1.2
ipython==8.22.2
iso8601==2.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
keras==3.0.5
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.6
lightgbm==4.0.0
lightning-utilities==0.10.0
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mock==5.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.2.4
namex==0.0.7
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==1.3.5
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
Pillow==10.0.0
pluggy==1.5.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PythonTurtle==0.3.2
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.25.3
rich==13.7.1
safetensors==0.4.3
schedule==1.2.1
scikit-image==0.23.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.0
selenium==4.15.2
serial==0.0.97
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2024.5.3
tokenizers==0.19.1
torch==2.1.1
torchmetrics==1.2.0
tqdm==4.63.2
traitlets==5.14.2
transformers==4.40.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
undetected-chromedriver==3.5.4
urllib3==2.0.3
uvicorn==0.30.1
uvloop==0.19.0
uWSGI==2.0.21
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xgboost==1.7.6
yarl==1.9.3
